- Removed update user activity cehck action from the stages to it's own controller - Added user authentication and authorization for the user activity check action
