+++ app/policies/user_activity_policy.rb
@@ -1,0 +1,11 @@
+class UserActivityPolicy < ApplicationPolicy
+  def update?
+    user.id == record.user_id
+  end
+
+  class Scope < Scope
+    def resolve
+      scope
+    end
+  end
+end
